update interface types
